requirements.txt
# Python 版本要求: >=3.8, <=3.12
# 核心依赖
volcengine-python-sdk[ark]==4.0.39
python-dotenv==1.0.1
openpyxl==3.1.5
